% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi2.R
\name{convert_to_minutes}
\alias{convert_to_minutes}
\title{Convert extracted latency values to minutesfor PSQI2}
\usage{
convert_to_minutes(data, start_col, end_col, unit_col)
}
\arguments{
\item{data}{A data frame}

\item{start_col}{Unquoted name of the column containing the start value}

\item{end_col}{Unquoted name of the column containing the end value}

\item{unit_col}{Unquoted name of the column indicating the time unit ("hr", "min", or "both")}
}
\value{
A data frame with start and end columns converted to minutes
}
\description{
Takes numeric start and end values and converts them to minutes based on the unit column.
Supports values expressed in hours, minutes, or both.
}
