% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi4.R
\name{finalize_hours}
\alias{finalize_hours}
\title{Calculate final total sleep hours value for PSQI4}
\usage{
finalize_hours(data, final_col, start_col, end_col)
}
\arguments{
\item{data}{A data frame}

\item{final_col}{Unquoted name of the column to create for the final sleep value}

\item{start_col}{Unquoted column name containing the start value (in hours)}

\item{end_col}{Unquoted column name containing the end value (in hours)}
}
\value{
A data frame with a new column containing the final total sleep hours value
}
\description{
Computes the final total sleep duration value for PSQI scoring. If both start and end values are present,
returns their midpoint. If only the start value is available, uses that. Values greater than 20 are set to NA.
}
