% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqi2.R
\name{finalize_latency}
\alias{finalize_latency}
\title{Calculate final PSQI latency value from start and end values for PSQI2}
\usage{
finalize_latency(data, final_col, start_col, end_col)
}
\arguments{
\item{data}{A data frame}

\item{final_col}{Unquoted name of the column to create for the final latency value}

\item{start_col}{Unquoted name of the column containing the start value (in minutes)}

\item{end_col}{Unquoted name of the column containing the end value (in minutes)}
}
\value{
A data frame with a new column containing the final latency value
}
\description{
Computes the final latency value for PSQI scoring. If both start and end values are present,
calculates their midpoint. If only the start value is present, uses that. Values of 400 or more are set to NA.
}
